requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy>=2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
python-multipart==0.0.9
